Move vertical and horizontal report tests to core tests
