+++ XReports.sln
@@ -54,6 +54,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XReports.Benchmarks.Console
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XReports.Tests.Analyzers", "tests\XReports.Tests.Analyzers\XReports.Tests.Analyzers.csproj", "{D4D39377-AD89-4265-B9C2-8F105B5D7081}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XReports.Tests.Common", "tests\XReports.Tests.Common\XReports.Tests.Common.csproj", "{8DC2DB68-61B0-4ACA-BC5A-8DFFC560B1BB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -104,6 +106,10 @@ Global
 		{D4D39377-AD89-4265-B9C2-8F105B5D7081}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D4D39377-AD89-4265-B9C2-8F105B5D7081}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D4D39377-AD89-4265-B9C2-8F105B5D7081}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8DC2DB68-61B0-4ACA-BC5A-8DFFC560B1BB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8DC2DB68-61B0-4ACA-BC5A-8DFFC560B1BB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8DC2DB68-61B0-4ACA-BC5A-8DFFC560B1BB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8DC2DB68-61B0-4ACA-BC5A-8DFFC560B1BB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{45EFC336-0978-4C5C-91B6-22541C463736} = {F9E94722-061B-4A10-896F-880442F76D1F}
@@ -119,5 +125,6 @@ Global
 		{0B54ECED-076D-4A2A-BC10-93FD160A105E} = {6E0C7C3A-81C6-453E-8FF9-1C3D96B97B88}
 		{43F53236-7373-4BE4-A846-3C0872C97C05} = {6E0C7C3A-81C6-453E-8FF9-1C3D96B97B88}
 		{D4D39377-AD89-4265-B9C2-8F105B5D7081} = {F9E94722-061B-4A10-896F-880442F76D1F}
+		{8DC2DB68-61B0-4ACA-BC5A-8DFFC560B1BB} = {F9E94722-061B-4A10-896F-880442F76D1F}
 	EndGlobalSection
 EndGlobal
