+++ tests/XReports.Core.Tests/SchemaBuilders/HorizontalReportSchemaBuilderTest.Basic.cs
@@ -4,12 +4,12 @@ using XReports.Extensions;
 using XReports.Interfaces;
 using XReports.Models;
 using XReports.SchemaBuilders;
-using XReports.Tests.Assertions;
+using XReports.Tests.Common.Assertions;
 using Xunit;
 
-namespace XReports.Tests.SchemaBuilders
+namespace XReports.Core.Tests.SchemaBuilders
 {
-    public partial class HorizontalReportTest
+    public partial class HorizontalReportSchemaBuilderTest
     {
         [Fact]
         public void BuildShouldSupportMultipleRows()
