+++ tests/XReports.Core.Tests/SchemaBuilders/HorizontalReportSchemaBuilderTest.TitleProperties.cs
@@ -3,12 +3,12 @@ using XReports.Extensions;
 using XReports.Interfaces;
 using XReports.Models;
 using XReports.SchemaBuilders;
-using XReports.Tests.Assertions;
+using XReports.Tests.Common.Assertions;
 using Xunit;
 
-namespace XReports.Tests.SchemaBuilders
+namespace XReports.Core.Tests.SchemaBuilders
 {
-    public partial class HorizontalReportTest
+    public partial class HorizontalReportSchemaBuilderTest
     {
         [Fact]
         public void BuildShouldSupportCustomHeaderProperty()
