+++ tests/XReports.Core.Tests/SchemaBuilders/HorizontalReportSchemaBuilderTest.cs
@@ -1,0 +1,6 @@
+namespace XReports.Core.Tests.SchemaBuilders
+{
+    public partial class HorizontalReportSchemaBuilderTest
+    {
+    }
+}
