+++ tests/XReports.Core.Tests/SchemaBuilders/VerticalReportSchemaBuilderTest.ComplexHeader.cs
@@ -2,13 +2,13 @@ using XReports.Extensions;
 using XReports.Interfaces;
 using XReports.Models;
 using XReports.SchemaBuilders;
-using XReports.Tests.Assertions;
+using XReports.Tests.Common.Assertions;
 using XReports.ValueProviders;
 using Xunit;
 
-namespace XReports.Tests.SchemaBuilders
+namespace XReports.Core.Tests.SchemaBuilders
 {
-    public partial class VerticalReportTest
+    public partial class VerticalReportSchemaBuilderTest
     {
         [Fact]
         public void BuildShouldSupportOneComplexHeaderRow()
