+++ tests/XReports.Tests/SchemaBuilders/AttributeBasedBuilderTest.BuildVerticalReport.cs
@@ -4,7 +4,7 @@ using XReports.Attributes;
 using XReports.Interfaces;
 using XReports.Models;
 using XReports.SchemaBuilders;
-using XReports.Tests.Assertions;
+using XReports.Tests.Common.Assertions;
 using Xunit;
 
 namespace XReports.Tests.SchemaBuilders
