+++ tests/XReports.Tests/SchemaBuilders/AttributeBasedBuilderTest.GlobalProperties.cs
@@ -5,7 +5,7 @@ using XReports.Interfaces;
 using XReports.Models;
 using XReports.Properties;
 using XReports.SchemaBuilders;
-using XReports.Tests.Assertions;
+using XReports.Tests.Common.Assertions;
 using Xunit;
 
 namespace XReports.Tests.SchemaBuilders
