--- tests/XReports.Tests/SchemaBuilders/VerticalReportTest.cs
@@ -1,6 +1,0 @@
-namespace XReports.Tests.SchemaBuilders
-{
-    public partial class VerticalReportTest
-    {
-    }
-}
